# app/api/api.yaml
openapi: 3.0.0
info:
  title: Koderpedia ERP API
  version: 1.0.0
servers:
  - url: 'http://localhost:1323'
    description: Local development server
...
paths:
...
  /api/v1/auth/user-api-contracts:
    get:
      tags:
        - Auth Access
      operationId: ListAuthUserApiContracts
      summary: List user API contract grants
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: User API contract grants listed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuthUserApiContractsResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    post:
      tags:
        - Auth Access
      operationId: CreateAuthUserApiContract
      summary: Create user API contract grant
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAuthUserApiContractRequest'
      responses:
        '201':
          description: User API contract grant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserApiContract'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
  '/api/v1/auth/user-api-contracts/{id}':
    parameters:
      - name: id
        in: path
        required: true
        description: User API contract grant id
        schema:
          type: string
          example: 1d7e2f23-b4c3-4ad7-8478-e770b68e6f11
    get:
      tags:
        - Auth Access
      operationId: GetAuthUserApiContract
      summary: Get user API contract grant
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: User API contract grant returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserApiContract'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    put:
      tags:
        - Auth Access
      operationId: UpdateAuthUserApiContract
      summary: Update user API contract grant
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAuthUserApiContractRequest'
      responses:
        '200':
          description: User API contract grant updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserApiContract'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    delete:
      tags:
        - Auth Access
      operationId: DeleteAuthUserApiContract
      summary: Delete user API contract grant
      description: Requires Authorization header and a user API contract grant.
      responses:
        '200':
          description: User API contract grant deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAuthAccessResponse'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
...
  '/api/v1/auth/users/{userId}/role':
    parameters:
      - name: userId
        in: path
        required: true
        description: Auth user id
        schema:
          type: string
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
    put:
      tags:
        - Auth Roles
      operationId: AssignAuthUserRole
      summary: Assign role to auth user
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignAuthUserRoleRequest'
      responses:
        '200':
          description: User role assigned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignAuthUserRoleResponse'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
    delete:
      tags:
        - Auth Roles
      operationId: DeleteAuthUserRole
      summary: Remove role from auth user
      description: Requires Authorization header and a user API contract grant.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - createdBy
              properties:
                createdBy:
                  type: string
                  description: Creator identifier
                  example: admin-user
      responses:
        '200':
          description: User role removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteAuthAccessResponse'
        '400':
          description: Bad request error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '401':
          description: 'Unauthorized, missing or invalid bearer token'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorized'
        '403':
          description: 'Forbidden, user has no API contract grant'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbidden'
components:
  schemas:
...
    ErrorBadRequest:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message describing the bad request
          example: Invalid input data
    ErrorUnauthorized:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message describing the unauthorized access
          example: 'Unauthorized, invalid OTP code'
    ErrorForbidden:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message describing the forbidden access
          example: 'Forbidden, you do not have permission to access this resource'
...
    AuthUserApiContract:
      type: object
      required:
        - id
        - userId
        - apiContractId
        - createdAtUtc0
        - updatedAtUtc0
        - isActive
      properties:
        id:
          type: string
          description: User API contract grant id
          example: 1d7e2f23-b4c3-4ad7-8478-e770b68e6f11
        userId:
          type: string
          description: Auth user id that receives API access
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        apiContractId:
          type: string
          description: API contract id granted to the user
          example: CryptoEncryptText
        createdAtUtc0:
          type: integer
          format: int64
          description: Created at in UTC+0 Unix millisecond time
          example: 1798790400000
        updatedAtUtc0:
          type: integer
          format: int64
          description: Updated at in UTC+0 Unix millisecond time
          example: 1798790400000
        isActive:
          type: boolean
          description: Whether the user API contract grant is active
          example: true
    CreateAuthUserApiContractRequest:
      type: object
      required:
        - userId
        - apiContractId
      properties:
        userId:
          type: string
          description: Auth user id that receives API access
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        apiContractId:
          type: string
          description: API contract id granted to the user
          example: CryptoEncryptText
    UpdateAuthUserApiContractRequest:
      type: object
      required:
        - userId
        - apiContractId
        - isActive
      properties:
        userId:
          type: string
          description: Auth user id that receives API access
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        apiContractId:
          type: string
          description: API contract id granted to the user
          example: CryptoEncryptText
        isActive:
          type: boolean
          description: Whether the user API contract grant is active
          example: true
    ListAuthUserApiContractsResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          description: Active user API contract grants
          items:
            $ref: '#/components/schemas/AuthUserApiContract'
          example:
            - id: 1d7e2f23-b4c3-4ad7-8478-e770b68e6f11
              userId: 358cbaad-316e-4539-9949-2636cdbd7e89
              apiContractId: CryptoEncryptText
              createdAtUtc0: 1798790400000
              updatedAtUtc0: 1798790400000
              isActive: true
    DeleteAuthAccessResponse:
      type: object
      required:
        - isSuccess
      properties:
        isSuccess:
          type: boolean
          description: Whether the delete operation succeeded
          example: true
...
    AssignAuthUserRoleRequest:
      type: object
      required:
        - roleId
        - createdBy
      properties:
        roleId:
          type: string
          description: Auth role id assigned to the user
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user
    AssignAuthUserRoleResponse:
      type: object
      required:
        - userId
        - roleId
        - roleName
        - grantedCount
        - updatedAtUtc0
        - createdBy
      properties:
        userId:
          type: string
          description: Auth user id assigned to the role
          example: 358cbaad-316e-4539-9949-2636cdbd7e89
        roleId:
          type: string
          description: Auth role id assigned to the user
          example: 1f2f4181-b4b1-4812-9f57-a6a315654040
        roleName:
          type: string
          description: Assigned role name
          example: admin
        grantedCount:
          type: integer
          format: int64
          description: Number of API contracts granted to the user from the role
          example: 12
        updatedAtUtc0:
          type: integer
          format: int64
          description: Updated at in UTC+0 Unix millisecond time
          example: 1798790400000
        createdBy:
          type: string
          description: Creator identifier
          example: admin-user